+++ upload/catalog/language/en-gb/extension/payhub/payment/payhub.php
@@ -1,0 +1,5 @@
+<?php
+$_['heading_title'] = 'PayHub';
+$_['text_title']    = 'Pay with PayHub';
+$_['text_failed']   = 'Payment failed.';
+$_['text_pending']  = 'Awaiting payment confirmation.';